[dy] Update models file
