+++ mage_ai/api/middleware.py
@@ -2,7 +2,7 @@ from json.decoder import JSONDecodeError
 from mage_ai.api.errors import ApiError
 from mage_ai.api.utils import authenticate_client_and_token
 from mage_ai.authentication.oauth2 import decode_token
-from mage_ai.orchestration.db.models import Oauth2Application
+from mage_ai.orchestration.db.models.oauth import Oauth2Application
 from mage_ai.settings import OAUTH2_APPLICATION_CLIENT_ID, REQUIRE_USER_AUTHENTICATION
 from mage_ai.server.api.constants import (
     COOKIE_OAUTH_TOKEN,
