+++ mage_ai/api/resources/PipelineResource.py
@@ -4,7 +4,7 @@ from mage_ai.data_preparation.models.constants import PipelineStatus
 from mage_ai.data_preparation.models.pipeline import Pipeline
 from mage_ai.data_preparation.repo_manager import get_repo_path
 from mage_ai.orchestration.db import safe_db_query
-from mage_ai.orchestration.db.models import PipelineSchedule, PipelineRun
+from mage_ai.orchestration.db.models.schedules import PipelineSchedule, PipelineRun
 from mage_ai.orchestration.pipeline_scheduler import PipelineScheduler
 from mage_ai.server.active_kernel import switch_active_kernel
 from mage_ai.server.kernels import PIPELINE_TO_KERNEL_NAME
