+++ mage_ai/orchestration/db/models/oauth.py
@@ -1,0 +1,98 @@
+from datetime import datetime
+from mage_ai.orchestration.db import safe_db_query
+from mage_ai.orchestration.db.errors import ValidationError
+from mage_ai.orchestration.db.models.base import BaseModel
+from sqlalchemy import (
+    Column,
+    Boolean,
+    DateTime,
+    Enum,
+    ForeignKey,
+    Integer,
+    String,
+)
+from sqlalchemy.orm import relationship, validates
+import enum
+import re
+
+
+class User(BaseModel):
+    avatar = Column(String(255), default=None)
+    email = Column(String(255), default=None, index=True, unique=True)
+    first_name = Column(String(255), default=None)
+    last_name = Column(String(255), default=None)
+    owner = Column(Boolean, default=False)
+    password_hash = Column(String(255), default=None)
+    password_salt = Column(String(255), default=None)
+    roles = Column(Integer, default=None)
+    username = Column(String(255), default=None, index=True, unique=True)
+
+    oauth2_applications = relationship('Oauth2Application', back_populates='user')
+    oauth2_access_tokens = relationship('Oauth2AccessToken', back_populates='user')
+
+    @validates('email')
+    def validate_email(self, key, value):
+        if value:
+            regex = re.compile(r"([-!#-'*+/-9=?A-Z^-~]+(\.[-!#-'*+/-9=?A-Z^-~]+)*|\"([]!#-[^-~ \t]|(\\[\t -~]))+\")@([-!#-'*+/-9=?A-Z^-~]+(\.[-!#-'*+/-9=?A-Z^-~]+)*|\[[\t -Z^-~]*])")  # noqa: E501
+            if not re.fullmatch(regex, value):
+                raise ValidationError('Email address is invalid.', metadata=dict(
+                    key=key,
+                    value=value,
+                ))
+        return value
+
+    @property
+    def roles_display(self) -> str:
+        if self.owner:
+            return 'Owner'
+        elif self.roles:
+            if self.roles & 1 != 0:
+                return 'Admin'
+            elif self.roles & 2 != 0:
+                return 'Editor'
+            elif self.roles & 4 != 0:
+                return 'Viewer'
+
+
+class Oauth2Application(BaseModel):
+    class AuthorizationGrantType(str, enum.Enum):
+        AUTHORIZATION_CODE = 'authorization-code'
+        CLIENT_CREDENTIALS = 'client-credentials'
+
+    class ClientType(str, enum.Enum):
+        PRIVATE = 'private'
+        PUBLIC = 'public'
+
+    authorization_grant_type = Column(
+        Enum(AuthorizationGrantType),
+        default=AuthorizationGrantType.AUTHORIZATION_CODE,
+    )
+    client_id = Column(String(255), index=True, unique=True)
+    client_type = Column(Enum(ClientType), default=ClientType.PRIVATE)
+    name = Column(String(255))
+    redirect_uris = Column(String(255), default=None)
+    user = relationship(User, back_populates='oauth2_applications')
+    user_id = Column(Integer, ForeignKey('user.id'))
+
+    oauth2_access_tokens = relationship('Oauth2AccessToken', back_populates='oauth2_application')
+
+    @classmethod
+    @safe_db_query
+    def query_client(self, api_key: str):
+        return self.query.filter(
+            Oauth2Application.client_id == api_key,
+        ).first()
+
+
+class Oauth2AccessToken(BaseModel):
+    expires = Column(DateTime(timezone=True))
+    oauth2_application = relationship(Oauth2Application, back_populates='oauth2_access_tokens')
+    oauth2_application_id = Column(Integer, ForeignKey('oauth2_application.id'))
+    token = Column(String(255), index=True, unique=True)
+    user = relationship(User, back_populates='oauth2_access_tokens')
+    user_id = Column(Integer, ForeignKey('user.id'))
+
+    def is_valid(self) -> bool:
+        return self.token and \
+            self.expires and \
+            self.expires >= datetime.utcnow().replace(tzinfo=self.expires.tzinfo)
