+++ mage_ai/orchestration/run_status_checker.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta
 from mage_ai.data_preparation.models.pipeline import Pipeline
 from mage_ai.data_preparation.repo_manager import get_repo_path
-from mage_ai.orchestration.db.models import BlockRun, PipelineRun, PipelineSchedule
+from mage_ai.orchestration.db.models.schedules import BlockRun, PipelineRun, PipelineSchedule
 from mage_ai.shared.array import find
 
 PIPELINE_FAILURE_STATUSES = [
