+++ mage_ai/tests/orchestration/test_pipeline_scheduler.py
@@ -1,7 +1,7 @@
 from mage_ai.data_preparation.models.constants import PipelineType
 from mage_ai.data_preparation.variable_manager import VariableManager
 from mage_ai.data_preparation.models.block import Block
-from mage_ai.orchestration.db.models import BlockRun, PipelineRun
+from mage_ai.orchestration.db.models.schedules import BlockRun, PipelineRun
 from mage_ai.orchestration.job_manager import JobType
 from mage_ai.orchestration.pipeline_scheduler import PipelineScheduler
 from mage_ai.tests.base_test import DBTestCase
